deepfake_detection_py/requirements.txt
torch
torchvision
timm
peft
transformers
Pillow
numpy
opencv-python
tqdm
matplotlib
facenet-pytorch
fastapi
uvicorn
python-multipart
scikit-learn
timesformer-pytorch
albumentations
requests
yt-dlp
